show_[czn|pco]_gpios - show platform iomux & GPIO info
These two scripts print out a table of the enabled iomux & GPIO settings
on a board. They read the register values and show the meaning of what's
enabled.

Unfortunately, on CZN and newer chips, a number of registers, including
the iomux registers may be locked down for security purposes. This means
that the registers cannot be read or written from the X86.

TODO: CZN's drive strengths vary based on whether the GPIO is 1.8 or 3.3
volts. A table of the voltage for each GPIO should be added, and the
drive strength should be updated for 3.3V GPIOs.

Signed-off-by: Martin Roth <gaumless@gmail.com>